Unify druid-shell RunLoop and Application types
