+++ druid-shell/src/platform/gtk/mod.rs
@@ -20,6 +20,5 @@ pub mod dialog;
 pub mod error;
 pub mod keycodes;
 pub mod menu;
-pub mod runloop;
 pub mod util;
 pub mod window;
