+++ druid-shell/src/platform/mac/mod.rs
@@ -22,6 +22,5 @@ pub mod dialog;
 pub mod error;
 pub mod keycodes;
 pub mod menu;
-pub mod runloop;
 pub mod util;
 pub mod window;
